+++ src/apis/config/publicApi.ts
@@ -1,7 +1,7 @@
 import axios from 'axios';
 
 import { CustomInstance } from './type';
-import { BASE_API_URL } from '@/constants/env';
+import { BASE_API_URL } from '@/constants';
 
 const publicApi: CustomInstance = axios.create({
   baseURL: `${BASE_API_URL}/api`,
