+++ src/app/(Main)/pick/PickComponent.tsx
@@ -9,7 +9,7 @@ import { usePosePickQuery } from '@/apis';
 import { BottomFixedDiv, PrimaryButton } from '@/components/Button';
 import PoseImage from '@/components/Modal/PoseImage';
 import { SelectionBasic } from '@/components/Selection';
-import { peopleCountList } from '@/constants/data';
+import { peopleCountList } from '@/constants';
 import useDidMount from '@/hooks/useDidMount';
 
 const DEFAULT_IMAGE = '/images/image-frame.png';
