+++ src/app/(Sub)/menu/withdraw/page.tsx
@@ -7,7 +7,7 @@ import { BottomFixedDiv, PrimaryButton } from '@/components/Button';
 import Header from '@/components/Header';
 import { Popup } from '@/components/Modal';
 import { useOverlay } from '@/components/Overlay/useOverlay';
-import { withdrawReasonList } from '@/constants/data';
+import { withdrawReasonList } from '@/constants';
 
 const RadioInput = ({ checked }: { checked: boolean }) => {
   return checked ? (
