+++ src/utils/isProduction.ts
@@ -1,3 +1,3 @@
-import { NODE_ENV } from '@/constants/env';
+import { NODE_ENV } from '@/constants';
 
 export const isProduction = NODE_ENV === 'production';
